Bug fixes and cosmetic changes (#408)
* partially fixing #375

* adding __len__ to df

* passing pycylon.Table to df operations

* fixing buffer copy bug

* logging status errors in macros

* bug fix

* test case for chunked tables

* cosmetic changes

+++ cpp/src/cylon/join/join_utils.cpp
@@ -181,7 +181,7 @@ arrow::Status build_final_table(const std::vector<int64_t> &left_indices,
 }
 
 arrow::Status CombineChunks(const std::shared_ptr<arrow::Table> &table,
-                            int64_t col_index,
+                            int col_index,
                             std::shared_ptr<arrow::Table> &output_table,
                             arrow::MemoryPool *memory_pool) {
   if (table->column(col_index)->num_chunks() > 1) {
